+++ src/agents/mod.rs
@@ -1,29 +1,28 @@
 pub mod errors;
-pub mod spo_agents;
+pub mod utils;
+// pub mod spo_agents;
 use crate::memory::MessageVector;
 use anyhow::anyhow;
 pub use errors::AgentError;
 use serde_json::Value;
-
-use super::agents::spo_agents::AgentObserver;
 
 use crate::{
     language_models::{
         openai::{functions::CustomFunction, gpt::streaming_utils::*},
         LanguageModel,
     },
-    memory::{Memory, ToMessage},
+    memory::{cache::Memory, Message},
 };
 
 /// Agent struct for interracting with LLM
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct Agent {
-    /// Memory handles how agent recalls and caches memory
+    /// Unique Identifier
+    pub id: String,
+    /// Memory contains cache and ltm
     pub memory: Memory,
     /// Language model defines which model to use for the given agent
     pub model: LanguageModel,
-    /// Another agent observes this one to enhance prompting and chain of thought
-    pub observer: Option<AgentObserver>,
 }
 
 impl Default for Agent {
@@ -31,12 +30,12 @@ impl Default for Agent {
     fn default() -> Self {
         let init_prompt = crate::persistance::prompts::get_prompt_by_name("DEFAULT_INIT_PROMPT")
             .unwrap_or(MessageVector::init());
-        let memory = Memory::build().init_prompt(init_prompt).finished();
+        let memory = Memory::from(init_prompt);
         let model = LanguageModel::default_gpt();
         Agent {
+            id: uuid::Uuid::new_v4().to_string(),
             memory,
             model,
-            observer: None,
         }
     }
 }
@@ -47,23 +46,23 @@ impl Agent {
     /// * Updates gpt token_count
     /// * returns response as a string
     #[tracing::instrument(name = "Prompt agent for response")]
-    pub async fn prompt(&mut self, input: impl ToMessage) -> Result<String, AgentError> {
-        match &mut self.observer {
-            Some(obs) => {
-                if obs.has_pre_prompt_protocol() {
-                    let new_input = obs
-                        .mutate_input(self.memory.cache(), &input.to_string())
-                        .await;
-                    self.memory
-                        .push_to_message_cache(Some("user"), new_input)
-                        .await;
-                }
-            }
-            None => {
-                self.memory.push_to_message_cache(Some("user"), input).await;
-            }
-        }
+    pub async fn prompt(&mut self, message: Message) -> Result<String, AgentError> {
+        // match &mut self.observer {
+        //     Some(obs) => {
+        //         if obs.has_pre_prompt_protocol() {
+        //             let new_input = obs
+        //                 .mutate_input(self.memory.cache(), &input.to_string())
+        //                 .await;
+        //             self.memory
+        //                 .push_to_message_cache(Some("user"), new_input)
+        //                 .await;
+        //         }
+        //     }
+        //     None => {
+        //     }
+        // }
 
+        self.memory.force_push_message_to_cache(message);
         let gpt = self.model.inner_mut_gpt().unwrap();
         let cache = self.memory.cache();
         let response = gpt
@@ -85,26 +84,26 @@ impl Agent {
             .parse()
             .map_err(|err| AgentError::Undefined(anyhow!("Error parsing Gpt Reponse: {err:?}")))?;
 
-        match &self.observer {
-            Some(obs) => if obs.has_post_prompt_protocol() {},
-            None => {
-                self.memory
-                    .push_to_message_cache(Some("assistant"), parsed_response.to_owned())
-                    .await;
-            }
-        }
+        // match &self.observer {
+        //     Some(obs) => if obs.has_post_prompt_protocol() {},
+        //     None => {
+        //         self.memory
+        //             .push_to_message_cache(Some("assistant"), parsed_response.to_owned())
+        //             .await;
+        //     }
+        // }
 
         Ok(parsed_response)
     }
 
     /// Openai function calling completion
-    #[tracing::instrument(name = "Function prompt GPT API for response" skip(input, custom_function))]
+    #[tracing::instrument(name = "Function prompt GPT API for response" skip(message, custom_function))]
     pub async fn function_prompt(
         &mut self,
         custom_function: CustomFunction,
-        input: impl ToMessage,
+        message: Message,
     ) -> Result<Value, AgentError> {
-        self.memory.push_to_message_cache(Some("user"), input).await;
+        self.memory.force_push_message_to_cache(message);
         let func = custom_function.function();
         let gpt = &self.model.inner_gpt().unwrap();
         let cache = self.memory.cache();
@@ -123,9 +122,9 @@ impl Agent {
     #[tracing::instrument(name = "Prompt agent for stream response")]
     pub async fn stream_prompt(
         &mut self,
-        input: impl ToMessage,
+        message: Message,
     ) -> Result<CompletionReceiverHandler, AgentError> {
-        self.memory.push_to_message_cache(Some("user"), input).await;
+        self.memory.force_push_message_to_cache(message);
         let gpt = &self.model.inner_gpt().unwrap();
         let cache = self.memory.cache();
         let response_stream = gpt.stream_completion(&cache.into()).await?;
