+++ src/agents/spo_agents/summarizer.rs
@@ -20,13 +20,13 @@ impl SummarizerAgent {
     pub fn agent(&self) -> Agent {
         match self {
             Self::General => {
-                let init_prompt = MessageVector::from_message(
-            r#"You are a code summarization Ai, you will be given a chunk of code to summarize
+                let init_prompt = MessageVector::new(
+                    r#"You are a code summarization Ai, you will be given a chunk of code to summarize
                 - Mistakes erode user's trust, so be as accurate and thorough as possible
                 - Be highly organized 
                 - Do not use lists or anything resembling a list in your summary
-                - think through your response step by step, your summary should be succinct but accurate"#
-        .to_string().to_message_with_role(MessageRole::System));
+                - think through your response step by step, your summary should be succinct but accurate"#,
+                );
 
                 let memory = Memory::build()
                     .init_prompt(init_prompt)
@@ -40,15 +40,13 @@ impl SummarizerAgent {
                 }
             }
             Self::Memory => {
-                let init_prompt = MessageVector::from_message(
+                let init_prompt = MessageVector::new(
                     r#"You are a memory summarizer agent, given the message thread, 
             your task is to provide a concise and informative summary of the 
             conversation. Consider the user's input and your responses. 
             Be accurate, organized, and include key points from the dialogue. 
             Your summaries should be helpful and reflect the essence of the 
-            conversation. Remember, clarity and brevity are essential."#
-                        .to_string()
-                        .to_message_with_role(MessageRole::System),
+            conversation. Remember, clarity and brevity are essential."#,
                 );
 
                 let memory = Memory::build()
