+++ src/agents/utils.rs
@@ -1,0 +1,1 @@
+pub const FILE_SUMMARIZER_SYS_PROMPT: &str = "You are a summarizer AI, you will be given the content of a file. Please create a succinct and detailed summary of what the code in the file does. Do not mention you're an AI.";
