+++ src/features/long_term_memory/database/.handlers/file.rs
