+++ src/features/long_term_memory/database/.handlers/threads.rs
