+++ src/features/long_term_memory/database/.models/file_chunks.rs
