+++ src/features/long_term_memory/database/mod.rs
@@ -1,8 +1,9 @@
-pub mod api;
+// pub mod api;
 pub mod handlers;
 pub mod init;
 pub mod models;
-pub mod vector_embeddings;
+// pub mod vector_embeddings;
+// pub mod sql_message;
 
-pub use api::*;
+// pub use api::*;
 pub use init::DbPool;
