--- src/features/long_term_memory/database/models/mod.rs
@@ -1,4 +1,0 @@
-pub mod file;
-pub mod file_chunks;
-pub mod messages;
-pub mod threads;
