+++ src/features/long_term_memory/sqlx_embeddings.rs
