+++ src/language_models/huggingface/mod.rs
@@ -1,1 +1,2 @@
 pub mod sentence_embeddings;
+pub use sentence_embeddings::embed;
