+++ src/memory/embeddings/mod.rs
@@ -1,5 +1,3 @@
-pub mod embedded_structs;
-pub mod sqlx_wrapper;
 pub mod vector_model;
 
 pub use vector_model::EmbeddingVector;
